Bubblesort

Description

functions

bubblesort :: Ord a => [a] -> [a]

Function which sorts elements in a table by using bubblesort algorithm

bubblesort_loop :: Ord a => [a] -> Int -> [a]

Function which applies function bubblesort_one_iteration n times, when n is a length of the table.

bubblesort_one_iteration :: Ord a => [a] -> [a]

Function which exchanges pair of elements in table if it is in a wrong order. Simply, one iteration of bubblesort algorithm.