Heapsort

Description

functions

parent :: (Integral a, Num a, Ord a) => a -> a

Function which takes index of a cell in a table and returns its parent

leftChild :: Num a => a -> a

Function which takes index of a cell in a table and returns its left child

rightChild :: Num a => a -> a

Function which takes index of a cell in a table and returns its right child

heapify :: Ord a => [a] -> Int -> [a]

Function which restores a heap condition in a node

buildHeap :: Ord a => [a] -> Int -> [a]

Function which creates a heap out of a table

heapSortLoop :: Ord a => [a] -> Int -> [a]

Function which is a part of sorting algorithm

heapsort :: Ord a => [a] -> [a]

Function which sorts elements in a table by using heapsort algorithm