Insertionsort

Description

functions

insertionsort :: Ord a => [a] -> [a]

Function which calls function insert with empty tail (sorted by default) at the beggining and unsorted list. In the end an output list is sorted.

insert :: Ord a => [a] -> [a] -> [a]

Function which recursively inserts a head of table to into appropriate place. Function is executed until there’s nothing left unsorted.