Mergesort

Description

functions

firststhalf :: [a] -> [a]

Function which returns first half of the list to sort.

secondhalf :: [a] -> [a]

Function which returns second half of the list to sort.

merge :: Ord a => [a] -> [a] -> [a]

Function which receives two sorted lists (firsthalf and secondhalf) and produces one merged list. Function compares the first elements of this two lists and creates sorted list.

mergesort :: Ord a => [a] -> [a]

Function which sorts elements in a list by using mergesort algorithm.