Selectionsort

Description

functions

selectionsort :: Ord a => [a] -> [a]

Function which sorts elements in a table by using selectionsort algorithm. SelectionSort executed on an empty list is an empty list. Otherwise it adds the smallest element from the list to the beginning of the sorted list.

delete :: Eq a => a -> [a] -> [a]

Function which deletes first occurance in list of given value. If a value which we want to delete is head, we simply return tail. Otherwise we search that value in tail.